// apps/server/src/events/events.types.ts
export enum EventDataActionType {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  ARCHIVED = 'ARCHIVED',
}

registerEnumType(EventDataActionType, {
  name: 'EventDataActionType',
  description: 'Available event data action types',
});

registerEnumType(EventDataActionType, {
  name: 'EventDataActionType',
  description: 'Available event data action types',
});

export enum EventChannel {
  WORKSPACE = 'WORKSPACE',
  PERSONAL = 'PERSONAL',
  NONE = 'NONE',
}